requirements.txt
requests>=2.31.0
flask>=3.0.0
google-generativeai>=0.7.0
gunicorn>=21.2.0
